+++ demo/src/main/java/com/example/demo/repository/ClienteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<ClientesEntity,Long> {
+public interface ClienteRepository extends JpaRepository<ClientesEntity,Integer> {
 }
